requirements.txt
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.15.0
cssselect2==0.4.1
defusedxml==0.7.1
pep8==1.7.1
Pillow==9.0.1
pycparser==2.21
tinycss2==1.1.1
webencodings==0.5.1
